Only pull termianl commands from public monobehaviour methods
