+++ Assets/OpenTerminal/Scripts/SampleCommand.cs
@@ -1,6 +1,6 @@
 using UnityEngine;
 
-public class SampleCommand
+public class SampleCommand : MonoBehaviour
 {
     [TerminalCommand("debug", "Debugs a sample line in Unity Console")]
     public void SampleDebug(string input)
